FIXED FRAME & ADAPT

+++ Class/CMenu.h
@@ -26,8 +26,8 @@ private:
 
 public:
 	virtual void Init(SDL_Surface * screen);
-	virtual void Update(float dt);
-	virtual void Render(float dt);
+	virtual void Update(int dt);
+	virtual void Render(int dt);
 	virtual void Exit();
 };
